chore: document and split ctxval package
