+++ internal/ctxval/keys.go
@@ -1,0 +1,11 @@
+// Provides context value operations.
+package ctxval
+
+type commonKeyId int
+
+const (
+	loggerCtxKey         commonKeyId = iota
+	requestIdCtxKey      commonKeyId = iota
+	accountIdCtxKey      commonKeyId = iota
+	unleashContextCtxKey commonKeyId = iota
+)
